Refactor Code
Added Unicode Support
SourcePath is default the fullPath
